+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.4.2 - 2021-03-03
-
+## 1.4.3 - 2021-3-30
 ### Changed
 - Primary color is default color for all labels in qas-tabs-generetor
 - Default class is text-bold for selected label in qas-tabs-generator 
+
+## 1.4.2 - 2021-03-03
 
 ### Added
 - added vue to dependencies (`/ui`)